Maj services

+++ src/main/java/tn/esprit/syrinetrabelsi4arctic3/Repos/SikierRepository.java
@@ -3,5 +3,5 @@ package tn.esprit.syrinetrabelsi4arctic3.Repos;
 import org.springframework.data.jpa.repository.JpaRepository;
 import tn.esprit.syrinetrabelsi4arctic3.Entity.Skier;
 
-public interface SikierRepository extends JpaRepository<Skier,Integer> {
+public interface SikierRepository extends JpaRepository<Skier,Long> {
 }
